% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnmf.R
\name{computeWeight}
\alias{computeWeight}
\title{Compute Weight Matrix}
\usage{
computeWeight(coil20$X, weightType="heat-kernel", sigma=0.5)
}
\arguments{
\item{input}{a matrix of input data.}

\item{weightType}{the method by which the weight matrix is computed.}

\item{sigma}{parameter used in heat-kernel weight matrix calculation, default 0.3.}
}
\value{
A matrix of weights
}
\description{
Computes a weight matrix for the input matrix provided using method
requested.
}
\examples{
coil20  <- readMat("./testData/COIL20.mat")
W  <- computeWeight(coil20$X, weightType="dot-weighting")

}
