% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnmf.R
\name{gnmfCai}
\alias{gnmfCai}
\title{GNMF based on Cai et al's implementation}
\usage{
gnmfCai(input, nClust, weightMatrix, alpha, minIter, normalizeW, maxIter=NA,
               diffError=1e-5, nRepeat=10, meanFitRatio=0.1)
}
\arguments{
\item{input}{a matrix of input data, must be positive.}

\item{nClust}{desired number of features to reduce to}

\item{weightMatrix}{weight matrix returned by computeWeight}

\item{alpha}{parameter that scales W.  IF this parameter is 0 or less this function
returns NMF.}

\item{minIter}{minimum number of iterations to complete}

\item{normalizeW}{boolean variable indicating whether or not W should be normalized}

\item{maxIter}{optional parameter, specifies the maximum number of iterations}

\item{diffError}{optional parameter, used as threshold against which error is
measured}

\item{nRepeat}{number of times to repeat the optimization, defaults to 10}

\item{meanFitRatio}{scales mean fit, defaults to 0.1}
}
\description{
implemention of GNMF in R and Rcpp that is based on Cai et al's
implementation in Matlab.
}
\examples{
uv <- gnmfCai(coil20$X, 8, W, alpha=1, minIter=5, normalizeW=TRUE)

}
